added granite_docling test in separate python file
